# Optimize setNode