--- a/.mocharc.yaml
+++ b/.mocharc.yaml
@@ -1,0 +1,2 @@
+colors: true
+require: ts-node/register