--- a/package.json
+++ b/package.json
@@ -10,7 +10,7 @@
     "check-types": "tsc --noEmit",
     "build": "tsc",
     "lint": "eslint --color --ext .ts src/",
-    "test": "mocha -r ts-node/register 'test/**/*.test.ts'"
+    "test": "mocha -r ts-node/register 'test/unit/**/*.test.ts'"
   },
   "pre-push": [
     "lint"
@@ -44,13 +44,14 @@
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-prettier": "^3.3.1",
     "karma": "^4.3.0",
-    "mocha": "^6.2.0",
+    "mocha": "^8.3.0",
     "nyc": "^14.1.1",
     "prettier": "^2.2.1",
-    "ts-node": "^9.1.0",
+    "ts-node": "^9.1.1",
     "typescript": "^4.1.3"
   },
   "dependencies": {
-    "@chainsafe/as-sha256": "^0.2.2"
+    "@chainsafe/as-sha256": "^0.2.2",
+    "@dapplion/benchmark": "^0.1.6"
   }
 }