--- a/package.json
+++ b/package.json
@@ -10,7 +10,8 @@
     "check-types": "tsc --noEmit",
     "build": "tsc",
     "lint": "eslint --color --ext .ts src/",
-    "test": "mocha -r ts-node/register 'test/unit/**/*.test.ts'"
+    "test": "mocha -r ts-node/register 'test/unit/**/*.test.ts'",
+    "benchmark": "node -r ts-node/register ./node_modules/.bin/benchmark 'test/perf/**/*.test.ts'"
   },
   "pre-push": [
     "lint"