--- a/package.json
+++ b/package.json
@@ -33,6 +33,7 @@
   },
   "homepage": "https://github.com/ChainSafe/persistent-merkle-tree#readme",
   "devDependencies": {
+    "@dapplion/benchmark": "^0.1.6",
     "@types/chai": "^4.2.0",
     "@types/mocha": "^5.2.7",
     "@types/node": "^14.14.0",
@@ -52,7 +53,6 @@
     "typescript": "^4.1.3"
   },
   "dependencies": {
-    "@chainsafe/as-sha256": "^0.2.2",
-    "@dapplion/benchmark": "^0.1.6"
+    "@chainsafe/as-sha256": "^0.2.2"
   }
 }