--- a/package.json
+++ b/package.json
@@ -12,7 +12,7 @@
     "lint": "eslint --color --ext .ts src/",
     "benchmark": "node --max-old-space-size=4096 --expose-gc -r ts-node/register ./node_modules/.bin/benchmark 'test/perf/*.perf.ts'",
     "benchmark:local": "yarn benchmark --local",
-    "test": "mocha -r ts-node/register 'test/**/*.test.ts'"
+    "test": "mocha -r ts-node/register 'test/unit/**/*.test.ts'"
   },
   "pre-push": [
     "lint"