--- a/test/unit/proof/treeOffset.test.ts
+++ b/test/unit/proof/treeOffset.test.ts
@@ -1,7 +1,7 @@
 import {expect} from "chai";
 import {describe, it} from "mocha";
-import {createNodeFromTreeOffsetProof, createTreeOffsetProof} from "../../src/proof/treeOffset";
-import {zeroNode} from "../../src/zeroNode";
+import {createNodeFromTreeOffsetProof, createTreeOffsetProof} from "../../../src/proof/treeOffset";
+import {zeroNode} from "../../../src/zeroNode";
 
 describe("computeTreeOffsetProof", () => {
   it("should properly compute known testcases", () => {