--- a/test/unit/tree.test.ts
+++ b/test/unit/tree.test.ts
@@ -61,16 +61,16 @@ describe("tree correctness", () => {
   it("Should compute root correctly after setting a leaf", () => {
     const depth = 4;
     const tree = new Tree(zeroNode(depth));
-    tree.setNode2(BigInt(18), new LeafNode(Buffer.alloc(32, 2)));
+    tree.setNode(BigInt(18), new LeafNode(Buffer.alloc(32, 2)));
     expect(toHex(tree.root)).to.equal("3cfd85690fdd88abcf22ca7acf45bb47835326ff3166d3c953d5a23263fea2b2");
   });
 
   it("Should compute root correctly after setting 3 leafs", () => {
     const depth = 5;
     const tree = new Tree(zeroNode(depth));
-    tree.setNode2(BigInt(18), new LeafNode(Buffer.alloc(32, 2)));
-    tree.setNode2(BigInt(46), new LeafNode(Buffer.alloc(32, 2)));
-    tree.setNode2(BigInt(60), new LeafNode(Buffer.alloc(32, 2)));
+    tree.setNode(BigInt(18), new LeafNode(Buffer.alloc(32, 2)));
+    tree.setNode(BigInt(46), new LeafNode(Buffer.alloc(32, 2)));
+    tree.setNode(BigInt(60), new LeafNode(Buffer.alloc(32, 2)));
     expect(toHex(tree.root)).to.equal("02607e58782c912e2f96f4ff9daf494d0d115e7c37e8c2b7ddce17213591151b");
   });
 });