**Motivation**

Keeping a record of the parent nodes in an array instead of binded functions is faster.

**Description**

Refactor setNode to:
- Don't use iterators. Calculate entire bitstring at once
- Keep a list of parentNodes, then iterate it backwards to replace nodes.

Overall the result is about x2 faster

```
  set
    ✓ set at depth 8                             1607717 ops/s    622.0000 ns/op   x0.556    1057567 runs   1.05 s
    ✓ set at depth 16                            1254705 ops/s    797.0000 ns/op   x0.449     967763 runs   1.04 s
    ✓ set at depth 32                           882612.5 ops/s    1.133000 us/op   x0.425     731535 runs   1.03 s
```